Move GUI to its own folder
to avoid clashing with other code.
